Web/fastapi_app/requirements.txt
fastapi
uvicorn
python-multipart
mediapipe>=0.10.0
protobuf>=4.21.0
scikit-learn
ultralytics
numpy
opencv-python
pillow
